feat: Add leader board and user score components
The React components to show the leader board for the game were created. It will show the user with the highest
to the lowest score. The score is the sum of the number of questions answered and created by the user.

Another minor were done to improve Sign-in and New question functionalities.